requirements.txt
attrs==19.1.0
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
flasgger==0.9.2
Flask==1.1.1
Flask-API==1.1
Flask-PyMongo==2.3.0
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
mistune==0.8.4
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
pymongo==3.8.0
pyrsistent==0.15.4
PyYAML==5.1.2
six==1.12.0
Werkzeug==0.15.5
